requirements.txt
python-binance
pandas
numpy
ta
